Fix Android browser crash on startup after upgrade to cr136
fixes brave/brave-browser#45816

This is cherry-pick of
https://source.chromium.org/chromium/chromium/src/+/d1a0fdad538d22ab6fc23aba0d9ff583d21cb371

Fix a crash that happens when the user removes an enabled flag

When the user enables a flag and then removes the flag from codebase,
Chrome Android will crash on startup because it assumes that all
previously enabled flags exist. This CL fixes this crash by just
moving on if a previously enabled flag does not exist anymore.

Bug: 409983752
Change-Id: I5d92662c0111d31b2fb6a843d2fa02ba0e310dd0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6455543