Ensure that `GetBraveServiceName`/`GetBraveAccountName` return correct values
`GetBraveServiceName` and `GetBraveAccountName` initialize and return a static
value based on a command-line flag. If these functions are called before the
command-line flag is set (as occurs now when running browser tests), then the
static value ends up initialized with the wrong service or account name.
Instead, we should use individual static variables to hold the values and return
the correct one at run-time.