[plaster] Initial prototype (#28389)
This is the initial prototype for 🩹 Plaster. This work is based on a
recenlty worked out proposal:

https://docs.google.com/document/d/1aPf7cMB5Pc_ssKFC_I4-mLHr-7D4lTqsvItSXY-WzQM/

Run this tool with:

```
tools/cr/plaster.py
```

The tool at the moment is in a very early prototyping stage, and there
are many tickets opened for it, to get it into a mature state. The
intent is to manage to deliver this in gradual steps, while acquiring
experience alongside the way.

- Plaster files.

These are TOML files, even though they look pretty similar to ini files.

Each `[[substitution]]` entry represents an individual regex operation to be
applied to a file. The format for plasters is something like this

```toml
[[substitution]]
description = ''
re_pattern = ''
replace = ''
re_flags = ''
count = 0
```

A user can list as many of these entries as necessary for each regex
that one desires to apply to a given source.

Resolves https://github.com/brave/brave-browser/issues/45051