[cr138] Remove DCHECK that is being triggered during printing
This DCHECK is triggering while printing (via the browser and browser tests).
The print preview panel appears but it's completely blank. No relevant errors
are logged and it's unclear why the node is determined to need a style recalc.
There is no obvious upstream or Brave-related change tied to this failure. In
the debugger, it appeared to be operating on a style element related to
`--google-grey-600` when the DCHECK was triggered, but my experimentation with
that style and other related styles led nowhere, unfortunately. This will most
likely require some deeper investigation outside of the scope of the rebase to
address.