Fix includes for `base::StringPrintf`
Upstream is doing a couple of IWYU fixes for `base::StringPrintf`, and
this will have an effect on our own uses of this function. This change
corrects all the includes we have.

This is a mechanical change, done with the following script.

```sh

remove_header_if_unused() {
    files=$(git grep -l "base/strings/stringprintf.h")
    for file in $files; do
        if ! git grep -qE "base::StringPrintf|base::StringAppend" "$file"; then
            sed -i '/base\/strings\/stringprintf.h/d' "$file"
            echo "Removed 'base/strings/stringprintf.h' from $file"
        fi
    done
}

add_header_if_needed() {
    files=$(git grep -lE "base::StringPrintf|base::StringAppend")
    for file in $files; do
        ../tools/add_header.py --header '"base/strings/stringprintf.h"' "$file"
    done
}

remove_header_if_unused
add_header_if_needed
```