[iOS] Select child tabs created by the DOM immediately
This change removes the delay before selecting a tab that is created by the DOM (via `window.open` or `target=_blank`) and ensures that `opener` on the `TabState` is only set for these types of DOM-created tabs. A side effect of this is that a web view is no longer guaranteed to be created when a tab is selected, so scroll view observation must also be done in `tabDidCreateWebView`. This change also updates the copy for a failed download to not display Apple Wallet Passes failure messages.