Fixed gn check error
```
ERROR at //extensions/common/constants.cc:13:11: Can't include this header from here.
          ^--------------------------
The target:
  //brave/extensions:common
is including a file from the target:
  //build:chromeos_buildflags

It's usually best to depend directly on the destination target.
In some cases, the destination target is considered a subcomponent
of an intermediate target. In this case, the intermediate target
should depend publicly on the destination to forward the ability
to include headers.

Dependency chain (there may also be others):
  //brave/extensions:common -->
  //base:base --[private]-->
  //base:rust_logger --[private]-->
  //build:chromeos_buildflags
```