[Leo] Click on temporary chat menu item should toggle the switch (#29419)
We were only flipping the toggle when the toggle element itself is
clicked, while the expected behavior is the entire menu item should be
toggling the switch if clicked.

This PR fixes the bug by adding the onClick handler for the menu item
with interactive property set to true so it won't dismiss the menu when
the item is clicked to match the behavior where the toggle itself is
clicked.