[presubmit] Check patch files have correct source (#29482)
A recent PR ended up introducing a patch file with a name that indicates
one source, but the actual patch header pointing to a different file.
This causes issues when running `npm run update_patches` as the tooling
tries to correct the file name.
(https://github.com/brave/brave-core/pull/28954)

This change adds a new presubmit check that makes sure the patchfile
name matches the the name for the source being patched.

Resolves https://github.com/brave/brave-browser/issues/46690