[l10n]: Add a system for generating WebUI string bindings and Typescript types for those functions. (#29172)
This PR updates Grit to output:

- A C++ header file with the WebUI strings
- A Typescript file with the generated enum to access the upstream strings.

As a proof of concept it also migrates one string from AiChat to use the new string bindings. I've experimented converting them all over and it works well, just was churning a lot with AiChat updates and approach changes so I'll do that as a follow up.

---------

Co-authored-by: Aleksei Khoroshilov <akhoroshilov@brave.com>