Try traversing up the directory tree to find mangled files
Sometimes mangled files are not located in a directory corresponding to
the source file's directory. Instead, they seem to be placed at a
directory corresponding to BUILD.gn file's directory.

So when failing to find a mangled file in the source file's directory,
try traversing up the directory tree until a mangled file is found.