Crash when clicking on shields (#29457)
Fix for the crash on opening the Shields control panel

The change was made in context of observing lots of crash reports on
the backtrace. All crashes were related with opening Shields control
panel and retrieving the size and coordinates of the Sields panel.
The idea was to get rid of finding active browser window by the
profile and storing of it, as a member across all Shields panel's
lifetime. Instead we retrieve the browser window only at the moment
we need it (with using another approach), and don't have to store it
as a member.

Resolves: https://github.com/brave/brave-browser/issues/46584

---------

Signed-off-by: Vadym Struts <vstruts@brave.com>