[CodeHealth] Remove `NoDestructor` use in `md_text_buttons.h` (#29579)
This PR removes the use of `NoDestructor` for this particular map with
button themes. The data itself is `constexpr`-compatible already,
however some of the values are produced through calls to
`color_utils::AlphaBlend`, which cannot be used in a `constexpr`
context. To correct this, this PR creates a couple of constants to store
the result of the `AlphaBlend` call.

This PR also adds a test to make sure that the result of the
`AlphaBlend` call is kept consistent, and that some failure is produced
whenever a value slides.

Resolves https://github.com/brave/brave-browser/issues/46863