[plaster] Adding presubmit check for 🩹 files
This change adds support to plaster for `.patchinfo` files, allowing a
more efficient way to make sure any given plaster file has should be
regenerated. This introduces the necessary supporting machinery to
permit a `plaster.py check` call, that is being hooked into our
presubmits.

This change aims to mitigate unwanted/accidental changes to patch files
for which a plaster file already exists.

- What does plaster do to .patchinfo files now

Plaster now generates the contents of a .patchinfo file, and based on
that content it determines if one of these files must be updated:
source, patch, or patchinfo. A new key is being added to the patchinfo
format, namely, `plaster`, which is used to store the path and checksum
for the plaster file for a given patch/source set. Files are only
updated on disk if the resulting new content doesn't match the hash
of the file saved in disk.

Resolves https://github.com/brave/brave-browser/issues/45058