AI Chat Tool definition and EngineConsumer inclusion
A Tool is a feature the LLM can "use" by sending a tool request with some input parameters specified.
This PR defines the Tool base class and determines where to send the array of Tools in use for a particular engine request: `EngineConsumer::GenerateAssistantResponse`.
When the conversation is deciding which tools to use, it will consult `Tool::IsSupportedByModel()`, `Tool::IsAgentTool()`, `Tool::IsContentAssociationRequired()` as well as knowledge about a static list of Tools.

When an engine responds with a tool use request, that request will be added to the events of a conversation history entry (via `ToolUseEvent` - not implemented in this PR).
The Tool class in this PR will be used to provide the result of that tool use request, via `Tool::UseTool()`.
We will then provide a response to the tool in the `ToolUseEvent`'s output field (not implemented in this PR).
The LLM will be called again providing an additional role=Tool conversation history entry.

In order to keep other submissions' size as small as possible, and to submit in parallel, this PR only contains the base definition of Tool and where Tool gets passed. It does not contain the tool uses, the related `ToolUseEvent` tool use and response conversation event, any Engine implementation which passes a list of tools to an API, or any UI representation of Tools or tool uses.

Subsequent submissions will be:
- Conversation API tool handling
- OAI Chat API tool handling
- ConversationHandler getting an agentic loop to perform the usage of the Tools and send the responses back to the engine
- UI representation for tool uses
- Tool implementations