AI Chat unit test comparisons use mojo::Equals and custom printers
Instead of manually comparing each property, we can lean on mojo::Equals which deep compares two instances of a mojom struct. This is preferable to the manual way because when properties are added we might forget to add them to the test equality comparison util. We do add a custom `PrintTo` for each struct so that when a test fails we can more easily see which property differed between the instances being compared.
This submission also fixes issues that have been exposed as a result of not comparing every field:
- `conversation.model_key` was being deleted in the database when the first entry was submitted. This does not present as a visible bug in the product since there is almost always a second entry submitted immediately afterwards, once the LLM has completed the response, which added the model_key back. In general I found it hard to read that the conversation.model_key was updated when a new conversation_entry is added, so I moved it to its own explicit function.