AI Chat Tool Use Conversation Events (#29643)
A Tool is a feature the LLM can "use" by sending a tool request with some input parameters specified.
This will eventually get added to the events of a conversation history entry (via `ToolUseEvent` added in this PR).
We will then provide a response to the tool in the ToolUseEvent's output field (not implemented in this PR).
The LLM will be called again providing an additional role=Tool conversation history entry (not implemented in this PR).

We put tool use requests and responses in the same object (`ToolUseEvent`) for ease of access and UI rendering. Otherwise there would be a lot of cases of searching history to find a relevant tool use response for a given tool use request if we kept them in assistant and user (or tool) `ConversationTurn` items.

In order to keep this submission size as small as possible, it only includes what's necessary to define the mojom structs, so that other PRs that use them can be created in parallel.

Subsequent submissions will be:
- Tool base class
- Conversation API support for sending defined tools, tool use requests, and tool use responses
- ConversationHandler getting an agentic loop to perform the usage of the Tools and send the responses back to the engine
- UI representation for tool uses
- Tool implementations