[canary] Fix B🚀 test failures on Windows (#29675)
This PR corrects all the test failures on Windows. Most of them are
related to file path separator, and similar things. One failure in
particular had to do with `subprocess.check_output` dropping `\r` from a
string when enconding to utf-8, which may be something to follow up on
plaster, to make sure saved patches match exactly what was produced by
the terminal.

It is not clear why tests have not been running for Windows all along,
however this is soon to change, as it is clear the windows run is taking
place with https://github.com/brave/brave-core/pull/29592.

Resolves https://github.com/brave/brave-browser/issues/47027