Improve Block Elements design to allow blocking elements behind picker (#29333)
Improve Block Elements design to allow blocking elements behind picker.

Added an ability to minimize the Element picker's card to resolve the
situation, when it pops up above the target element, which one the user
wants to block.
Added a new design for the Element picker card, which is Nala styled,
to follow the design line.

Resolves https://github.com/brave/brave-browser/issues/45104

---------

Signed-off-by: Vadym Struts <vstruts@brave.com>