Fixed crash from clicking private/tor window's profile avatar button
fix https://github.com/brave/brave-browser/issues/47073

In cr138, kEnableImprovedGuestProfileMenu is enabled by default but
we're not ready to enable it. Fixed disabling it by default.
As this feature is deleted and already handled in cr139 branch,
we don't need to another f/u for enabling it by default in the future.

TEST=BraveToolbarViewTest.ClickAvatarButtonTest