Fixed crash from clicking private/tor window's profile avatar button (#29723)
* Fixed crash from clicking private/tor window's profile avatar button

fix https://github.com/brave/brave-browser/issues/47073

In cr138, kEnableImprovedGuestProfileMenu is enabled by default but we're not ready to enable it.
We call SetProfileIdentityInfo() always from BraveIncognitoMenuView but upstream doesn't allow to call it
when kEnableImprovedGuestProfileMenu is enabled.
Fixed disabling it by default. As this feature is deleted and already handled in cr139 branch,
we don't need another f/u for enabling it by default in the future.

TEST=BraveToolbarViewTest.ClickAvatarButtonTest

* Revert "[cr138] Guest profile image and title containers now have no children"

This reverts commit 4aba9c0a6a44388cff5e7997a5fc51ccd99e821b.

Reverted as we disabled kEnableImprovedGuestProfileMenu by default.
This commit should be reverted again in cr139 as this feature is enabled
by default and removed.