[plaster] Adding presubmit check or 🩹 files
This PR is a follow up on
https://github.com/brave/brave-core/pull/29716.

This PR adds a presubmit check that calls `plaster.py check`, and
validates to make sure that any submitted plaster file is up-to-date
with its patches, and applying correctly to the Chromium tree.

There's still some work to be done regarding deleted plaster files, and
what to do with a present patch file in such a case, but this is being
tracked in a separate ticket.

Resolves https://github.com/brave/brave-browser/issues/45058