[canary] Fix `apply_patches` provided source relativeness
When patches fail to apply during an `apply_patches` run, a report is
produced with the patch files that failed, and in some cases, with the
path for the source that they patch file applies to. Brockit uses that
information when it is present, to refer to the source file. However,
there is an issue on how this path has been used, as the path provided
is relative to Chromium's root path.

This PR fixes a bug where the repo path was being concatenated with the
provided source path. This has never been an issue because most of the
time we are dealing with merge conflicts in Chromium. However, when it
comes to other submodules are involved, it is necessary to make sure
that the path provided by `apply_patches` for the source file, is made
relative to the path of the submodule the source belongs to.

This PR also removes some of the `Path` conversions in our tests, in
favour of using `as_posix()`, as string comparison produces better error
failure output than `Path` comparisons.

Resolves https://github.com/brave/brave-browser/issues/47102