[CodeHealth] `PRESUBMIT` check for `nogncheck`
This PR adds a presubmit check for `nogncheck` uses, to make sure we are
using this lint suppresion label in places where appropriate. In
particular, the need for this check has come up due to the wrong
assumption that `gn` somehow lacked the ability to understand `#ifdef`
conditionals for header inclusions, when in fact, the issue with `gn`
checks in such cases is that there is a buildflag for a certain
exclusion, but the same exclusion has also to be applied in gn files.

Resolves https://github.com/brave/brave-browser/issues/47269