[cr139] Removes unneeded DCHECK.
DCHECK(browser);
if (!browser) {
  return;
}

is advised against pattern. Should either be CHECK or if. In this case,
it seems `if` is appropriate.

The DCHECK was causing some upstream unit tests to crash because by the
time OnboardingTabHelper::MaybeCreateForWebContents would come back from
PostTaskAndReplyWithResult the test would already finish and the browser
would be gone.