[CodeHealth] Remove unused `string` statics in `LocalDataFilesService`
This PR removes the static fields in `LocalDataFilesService` which are
`std::string` globals specifically added to this class to provide custom
testing hooks. These globals are never used, and the function setting
custom values for these static strings is never called.

This change removes these unnecessary values, and enables
`-Wexit-time-destructors` for this target to prevent back sliding.

Resolves https://github.com/brave/brave-browser/issues/47342