[iOS] Update window protection visible status immediately (#29885)
A prior change made the scene not connect the browser until after the profile loaded, which means that by the time we create a WindowProtection instance and have access to the profile the app has now already finished launching. This change ensures that we check if we need to show the browser lock immediately upon  a `WindowProtection` being created.