Add feature flag to hide vertical tab strip completely when collapsed
When a new feature flag is enabled, the vertical tab strip will be
completely hidden when it is collapsed, rather than just showing a
narrow strip. This allows users to have a cleaner interface when they
prefer not to see the vertical tab strip at all.

This is implemented by adding a new feature flag for it,
`kBraveVerticalTabsHideWhenCollapsed`, and making minimum width of
vertical tab strip tab strip region view have 4px when it's collapsed.

4px is needed for users to expand the vertical tab strip again by
hovering over the collapsed area.