Remove 'Tab search button' from vertical tab strip
In favor of the tab search button in the toolbar view,
we decided to remove the tab search button from the
vertical tab strip.

This change removes the tab search button from the
vertical tab strip region view and updates the layout
accordingly. It also removes the associated test file
and updates the build configuration to exclude the
tab search button from the vertical tabs.