[canary][cr140] Fix 3way conflict parsing with whitespace warnigns
Recently, when running the lift for `140.0.7281.1..140.0.7283.1`,
brockit was listing one of the files as just `errors.`. This was
happening because the `git apply` command was producing a last line with
a warning, like this:

```
brave/patches/chrome-android-java-AndroidManifest.xml.patch:81: trailing whitespace.
        <!--
Applied patch to 'chrome/android/java/AndroidManifest.xml' with conflicts.
U chrome/android/java/AndroidManifest.xml
warning: 1 line adds whitespace errors.
```

This change adds a test to cover a case like this, and corrects how
brockit is scraping the `apply` output, making sure we just don't assume
the source path is always produced in the last line.

Resolves https://github.com/brave/brave-browser/issues/47453