[Containers] Show icons on Containers in context menus (#29926)
This change retrieves the icon for the Containers from preferences, 
and displays it in context menus. The icon is rasterized using 
static image resources(vector icons).

The main logic for rasterization is in `ContainersIconGenerator`,
which implements the rasterization interface of ImageModel::ImageGenerator.
In this class, image will be generated using custom gfx::CanvasImageSource.

`ContainerModel` will instantiate `ui::ImageModel` with this generator, and
each context menus will use this icon. For this, context menu's delegate should
provide device scale factor.

<img width="335" alt="image" src="https://github.com/user-attachments/assets/29af8d74-ac7c-467a-8ea3-229a1607c0b6" />

<!-- Add brave-browser issue below that this PR will resolve -->
Resolves https://github.com/brave/brave-browser/issues/47117