[RELAND][bedrock] `PlaylistSidePanelCoordinator` as a browser window feature
Reland Notes: This PR was reverted because
`SidePanelUtil::PopulateGlobalEntries` was CHECKing for the coordinator
pointer, however we do not create the coordinator for popups or desktop
PWAs. This reland corrects that, adds test to cover that case, and also
guard the creation of the coordinator behind the playlist feature flag.

This PR converts `PlaylistSidePanelCoordinator` to a browser window
feature. This means abandoning the use of `BrowserUserData` for it,
and applying the principle outlined for these types of components in
https://chromium.googlesource.com/chromium/src/+/main/docs/chrome_browser_design_principles.md

Additionally, this change does away with PlaylistSidePanelWebView, as
it was merely a think layer over the base class, providing a WeakPtr
member. The base class itself, SidePanelWebUIView, has a weak ptr
factory that we could expose if we really need that type of semantics,
however for the use in this coordinator, it doesn't seem to be the case
that we need weak ptr semantics when there is in fact an observer for
the views life time, which can easily be used to reset the pointer to
the view.

Most Browser* layer violations have been correct, although there should
be a follow up to entirely remove Browser* and similar uses of
GetBrowserForMigrationOnly, by properly correcting the access to
BrowserView in other places.

Resolves https://github.com/brave/brave-browser/issues/47604