[iOS] Add browser menu quick action count customization
This change introduces the ability to customize how many items appear as quick actions (grid buttons) in the browser menu, with more than 4 creating multiple rows of grid items. It also introduces a new "Reset to Default" button at the bottom of the customize screen to allow users to reset back to the stock menu.