[Customize Chrome Panel] Wire CustomizeColorSchemeModeHandler with brave dark mode utils (#29985)
As we're managing the color scheme mode in local state, 
we shouldn't depend on the upstream code, which uses 
profile preferences via ThemeService.

CustomizeColorSchemeModeHandler is handler for "Customize chrome panel"
web ui to set color scheme mode.

This change overrides the CustomizeColorSchemeModeHandler
with BraveCustomizeColorSchemeModeHandler, 
which uses Brave's dark mode utilities, which depends on
local state instead of profile preference.


<!-- Add brave-browser issue below that this PR will resolve -->
Resolves https://github.com/brave/brave-browser/issues/47496