[Vertical Tabs] Remove the delay for expanding vertical tabs when hovering over the region (#30044)
Remove the delay for expanding vertical tabs when hovering over the region

We had experimented with a delay for expanding vertical tabs, but it has been
decided to remove this delay for a more immediate user experience. This change
ensures that when the user hovers over the vertical tab strip region, it will
expand immediately without waiting for a specified time.