[json64] Make `base_unittests` a default suite
This change adds `base_unittests` to the list of default suits to be
built and run with brave. This is being done so we have full test
coverage over the default code paths
`base::Value`/`JSONReader`/JSONWriter`, as these classes are having
overrides added to them to support 64bit numbers, but we have to make
sure we are not breaking the default behaviour.

Resolves https://github.com/brave/brave-browser/issues/47643