AI Chat Tools ConversationAPI implementation (#29627)
* AI Chat Tools ConversationAPI implementation

Note that our internal representation of Tool use requests and responses is to group them in the same struct - ToolUseEvent, but the OAI style API represents them differently (likely to preserve history structure and simplify the API definition). We put them in the same object for ease of access and UI rendering. Otherwise there would be a lot of cases of searching history to find a relevant tool use response for a given tool use request.

In order to keep this submission size as small as possible, it only includes what's necessary to construct calls to the ConversationAPI. Subsequent submissions will be:
- ConversationHandler getting an agentic loop to perform the usage of the Tools and send the responses back to the engine
- UI representation for tool uses
- Tool implementations