Remove FOREGROUND_SERVICE_MEDIA_PROJECTION permission (#30115)
android.permission.FOREGROUND_SERVICE_MEDIA_PROJECTION was introduced
in C139 and currently is used only for desktop Android which we do not
build and publish. We need to remove it as Google rejects our app
update until we describe and provide a video why
android.permission.FOREGROUND_SERVICE_MEDIA_PROJECTION is used.
Resolves: https://github.com/brave/brave-browser/issues/47739