[Tree Tabs] Add a new flag to enable the Tree Tab feature (#30094)
Add a new flag to enable the Tree Tab feature

We're trying to add a new feature - the Tree Tab.
This feature should be developed behind a flag so that we don't break
existing functionality.

The flag is defined in `browser/about_flags.cc` and `browser/ui/tabs/features.cc`.

Resolves https://github.com/brave/brave-browser/issues/47694