[iOS] Add Static & Component Xcode build schemes (#30196)
This allows devs to use Component and Static build configurations in addition to the standard Debug & Release builds. It also removes a few redundant arguments passed into GN for iOS builds based on Component builds.

It also adds an option to the fastlane commands to override the scheme based on a passed in build type