Set Web view for AI chat side panel to focus on show (#30179) (uplift to 1.81.x). (#30201)
Set Web view for AI chat side panel to focus on show (#30179)

This PR sets focus to the AI chat side panel web view when it is shown.
This is done to ensure that the user can immediately interact with the
chat interface without needing to manually focus on it.

Technical Details:
- Modified `brave_side_panel_utils.cc` to include focus handling logic.
- Added a check to clear focus before requesting it, addressing a bug in focus handling.

Co-authored-by: Sangwoo Ko <sangwoo108@gmail.com>