[msan] Fix uninitialised read caused by `SetCookieInfo`
This override was doing a copy of the value passed in, `cookie_info`,
and then changing the value of one of the fields to `kHide`, in order to
make things work properly for the rest of the things done by the
override, and avoid a `CHECK`. The issue here is that `cookie_info` is a
`const` reference, and `cookie_info` have a copy constructor, so the
fields were being copied manually. This worked for a short time, until
other fields were added to this struct, with some being uninitialised.

This change removes the whole copying of this type, as that is
bug-prone, with new fields being added, and rather overrides
`SetThirdPartyCookiesInfo` to always bail out, as this is the actual
effect we are looking for by mutating the value of `cookie_info`.
`SetThirdPartyCookiesInfo` is only used by `SetCookieInfo` in fact.

Resolves https://github.com/brave/brave-browser/issues/47875