[ubsan] Add test filter support for ubsan builds (#30254)
This PR adds support for test exclusion for ubsan CI runs in particular.
When it comes to UBSan, it is not clear how stable this modality is, and
we have already come across failures that are upstream issues, so we
need the ability to suppress some of these tests with ubsan.

For details on the upstream issue we are trying to suppress check:
https://crbug.com/433075170

Resolves https://github.com/brave/brave-browser/issues/47888