[bots] Use `unsafe_buffers_patchs.txt` from upstream (#30279)
This PR removes the handling of
`//build/config/unsafe_buffers_paths.txt` in `brave-core` and just
patches the upstream file with our exclusions. Having our own copy of
this exclusion file was necessary at first, in order to provide
duplicated entries to handle `src/` path entries. This is not required
anymore with angle-bracket `chromium_src` inclusions.

This change also cleans up one of the static gn arg values in
`config.js`, as we work to move all these values out of it.

Resolves https://github.com/brave/brave-browser/issues/47916