Reload only active split tab (#30269)
Resolves brave/brave-browser#47896

Reload button reloads tabs that currently selected tabs.
With SideBySide, all split tabs(left & right) are added in selected tab when it's activated.
So, both split tabs are all reloaded. But we only want to reload active split tab.
It's difficult to touch selection model because it's related whole split tabs behavior.
To resolve this, inactive split tab is removed from selected tabs in chrome::ReloadInternal()
when only that split tabs is selected.

TEST=SplitViewWithTabDialogBrowserTest.SplitViewReloadTest