implement CashAddr decoding and encoding
The CashAddress format specification can be found here:
https://github.com/Bitcoin-ABC/bitcoin-abc/blob/master/doc/standards/cashaddr.md
It encodes the hash160 of a public key (p2pkh address) or the hash160 of a script (p2sh address).

This type of addresses are used on several blockchains with the only difference being the prefix. The "ecash:" prefix is used on the eCash mainnet.
So this code could be reused in the future for other chains with only a minor adjustment, not hardcoding the prefix.

The code was originally submitted in the Bitcoin ABC repository:
https://github.com/Bitcoin-ABC/bitcoin-abc/blob/master/src/cashaddr.cpp
https://github.com/Bitcoin-ABC/bitcoin-abc/blob/master/src/cashaddrenc.cpp

I later extracted it so it could be used standalone in other C++ projects, without any reliance on other Bitcoin ABC classes or structs:
https://github.com/PiRK/ecashaddrconv

This commits reuses the latter with only minor adjustments:
- conform to the Brave codestyle (cannot use 'throw' with exceptions disabled, Complex class/struct needs an explicit out-of-line constructor)
- address clang warnings (`-Wunsafe-buffer-usage`)
- drop the CashAddress to/from legacy bitcoin address conversion code