Enabled BrowserFullscreenShowBookmarkBarSplitView
In SideBySide, bookmarks bar is visible when one of split tab is
NTP to prevent bookmarks bar visibility change whenever changing
active tab in same split tabs. It's different behavior with our split
view. In our split view, bookmark bar is only visible when active split
tab is NTP. To align with upstream's SideBySide behavior, we should
follow in that way when SideBySide is enabled.

BrowserFullscreenShowBookmarkBarSplitView failed due to this difference.