[CodeHealth] Simplify/modernise `eth_requests.h` funcs (#30161)
This PR touches on the functions in `eth_requests.h` and adjacent code
for them. We are doing a few things:

 - Passing `std::string_view` rather than `const std::string&`.
 - Correcting the naming convention for these function.
 - Using `std::optional` where possible to signal presence, rather than
   `""`.

Additionally, this PR does away with
`GetUnstoppableDomainsProxyReaderContractAddress` and the map behind it.
There are few gains with this change around this function. First, we are
now using a `constexpr` set, and avoiding the string allocations that
were taking place. Second, the places where this data set was being used
have been reviewed, and the conversion to lower case is not needed
anymore, which removes the extra allocation for that.

Finally, `MultichainCalls::GetChains` is also being removed. The style
guide specifies that member functions must have some relation to the
data of a class, which wasn't the case here. Additonally, the vector
allocation is really not necessary for this case. This function was
being used only by `JsonRpcService`, and the set returned by `GetChains`
was at this moment a duplication of the keys of
`kUnstoppableDomainsProxyReaderContractAddresses`. This PR moves all
this code to be a discreet implementation detail of `JsonRpcService`, to
simplify further the association of this data. However, a static testing
function has been made available through `JsonRpcService` for testing
code.

Resolves https://github.com/brave/brave-browser/issues/47794