[ubsan] Add `is_ubsan` support to `.env` (#30311)
This is a follow up to https://github.com/brave/brave-core/pull/30254.

Support to test filter files specifically for ubsan runs has been added
in previous work, however this has not been picked up on CI, as it is
also necessary to provide the `--is_ubsan` flag to `npm run tests`.

This change provides `.env` support for this flag, so we can avoid the
question of making the flags for `npm run build` and `npm run test` to
match, which will make the CI changes smoother.

Additionally, this PR fixes the filter selection for ubsan, as it was
passing in the target arch as well, rather than only platform.

Resolves https://github.com/brave/brave-browser/issues/47969